Appointment Payment Flow
State of the appointment page needs to be maintained while navigation happens between Appointment and Payment page. For time being, this is maintained by passing as props to payment, bring them back via props again. This is very dirty solution. Will have to refactor it with redux or other state management.